RavenDB-17760 : minor modifications : - move event OnBeforeCommit from StorageEnvironment to DocumentStorage, - set OnBeforeCommit to UpdateBucketStatsTreeBeforeCommit from ShardedDocumentsStorage ctor - append OnBeforeCommit to LowLevelTransaction.OnBeforeCommit from DocumentsTransaction ctor - move all bucket stats handling from DocumentStorage to ShardedDocumentsStorage
