+++ src/Raven.Server/Documents/Schemas/Revisions.cs
@@ -1,4 +1,5 @@
 ﻿using Raven.Server.Documents.Revisions;
+using Raven.Server.Documents.Sharding;
 using Sparrow.Server;
 using Voron;
 using Voron.Data.Tables;
@@ -149,7 +150,7 @@ namespace Raven.Server.Documents.Schemas
             schema.DefineIndex(new TableSchema.DynamicKeyIndexDef
             {
                 GenerateKey = RevisionsStorage.GenerateBucketAndEtagIndexKeyForRevisions,
-                OnEntryChanged = DocumentsStorage.UpdateBucketStats,
+                OnEntryChanged = ShardedDocumentsStorage.UpdateBucketStats,
                 IsGlobal = true,
                 Name = RevisionsBucketAndEtagSlice
             });
