+++ src/Raven.Server/Documents/Schemas/TimeSeries.cs
@@ -1,4 +1,5 @@
-﻿using Raven.Server.Documents.TimeSeries;
+﻿using Raven.Server.Documents.Sharding;
+using Raven.Server.Documents.TimeSeries;
 using Sparrow.Server;
 using Voron;
 using Voron.Data.Tables;
@@ -81,7 +82,7 @@ namespace Raven.Server.Documents.Schemas
                 TimeSeriesSchemaBase60.DefineIndex(new TableSchema.DynamicKeyIndexDef
                 {
                     GenerateKey = TimeSeriesStorage.GenerateBucketAndEtagIndexKeyForTimeSeries,
-                    OnEntryChanged = DocumentsStorage.UpdateBucketStats,
+                    OnEntryChanged = ShardedDocumentsStorage.UpdateBucketStats,
                     IsGlobal = true,
                     Name = TimeSeriesBucketAndEtagSlice
                 });
